services/retrieval-api/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
numpy==2.2.3
scikit-learn==1.6.1
httpx==0.28.1
openai==1.65.4
chromadb==1.1.1
PyJWT==2.10.1
livekit-api==1.0.7
sentence-transformers>=3.0.0
torch>=2.2.0
